requirements.txt
mediapipe==0.10.14
opencv-python==4.9.0.80
numpy==1.26.4
pillow==10.3.0
pytesseract==0.3.10
matplotlib==3.8.4
pandas==2.2.2